% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCovs.R
\name{calculateCovs}
\alias{calculateCovs}
\title{Generates covaration matrix from each batch's GRN}
\usage{
calculateCovs(grns, par)
}
\arguments{
\item{par}{Logical indicating whether to run cov on all batches in parallel.
Default: F}
}
\description{
Calls base::cov on each of the GRNs and returns a list of
covariation matrices
}
