% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coRRectoR.R
\name{coRRectoR}
\alias{coRRectoR}
\title{coRRectoR wrapper function}
\usage{
coRRectoR(exps, neighbours = 20, par = F)
}
\arguments{
\item{exps}{List of raw expression matrices where each item in list is
an expression matrix where the columns are the samples in that batch and
the rows are the genes. Default: none}

\item{neighbours}{The number of nearest neighbours considered before the
intersection across genes is calculated. Default: 20}

\item{par}{Logical indicating whether to process the GRN inference and
covariation calculations for each batch in parallel. Default: F}
}
\description{
Performs all stages of coRRectoR pipeline by calling subsequent
functions. Pipeline is:
\enumerate{
\item Inference of GRNs
\item Calculation of covariation matrices
\item Calculation of intersection genes
\item Normalisation selected genes and unselected genes
}
}
