% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateGRNs.R
\name{generateGRNs}
\alias{generateGRNs}
\title{Generates GRN for each batch's expression matrix}
\usage{
generateGRNs(exps, par = F)
}
\arguments{
\item{par}{Logical indicating whether to run GRN inference on all batches
in parallel. Default: F}
}
\description{
Calls MEANIE3 on each of the batches and returns a list of
GRNs
}
