% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparebulkAnalyseR.R
\name{landingPanel}
\alias{landingPanel}
\title{Generate the landing page panel of the shiny app}
\value{
The UI and Server components of the shiny module, that can be used
within the UI and Server definitions of a shiny app.
}
\description{
These are the UI and server components of the landing page
panel of the shiny app. It is a modified version of the bulkAnalyseR landing page.
}
