% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEANIE3.R
\name{MEANIE3}
\alias{MEANIE3}
\title{Modified GENIE3}
\usage{
MEANIE3(exprMatrix)
}
\arguments{
\item{exprMatrix}{Expression matrix where columns refer to samples and rows
to genes}
}
\description{
Modified version of GENIE3 that translates each column by its
mean and scales by its standard deviation
}
