% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normaliseExps.R
\name{normaliseExps}
\alias{normaliseExps}
\title{Creates a normalised expression matrix of all samples}
\usage{
normaliseExps(exps, uniqueGenes)
}
\arguments{
\item{exps}{List of raw expression matrices where each item in list is
an expression matrix where the columns are the samples in that batch and
the rows are the genes. Default: none}

\item{uniqueGenes}{List of selected genes, as calculated by coRRectoR::intersectGenes.
Default: none}
}
\description{
Merges all expression matrices (from each batch) into one and then
normalises the selected genes using quantile normalisation before using POIBM
to correct the difference between the selected and unselected genes.
}
