% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparebulkAnalyseR.R
\name{preparebulkAnalyseR}
\alias{preparebulkAnalyseR}
\title{Generate all files required for an autonomous shiny app}
\usage{
preparebulkAnalyseR(
  corrected.output,
  metadata,
  uncorrected = NULL,
  add.batch.column = TRUE,
  batch.names = NULL,
  create.comparison.app = TRUE,
  create.corrected.app = TRUE,
  comparison.app.dir = "comparisonApp",
  comparison.app.title = "Comparison between uncorrected and corrected outputs",
  corrected.app.dir = "correctedApp",
  corrected.app.title = "Batch corrected bulkAnalyseR app",
  organism = NA,
  org.db = NA,
  modality = "RNA",
  theme = "flatly",
  panels.default = c("Landing", "SampleSelect", "QC", "GRN", "DE", "DEplot", "DEsummary",
    "Enrichment", "GRNenrichment", "Cross", "Patterns")
)
}
\arguments{
\item{metadata}{Either one metadata dataframe (combined across batches)
or multiple metadata dataframes in a list}

\item{uncorrected}{An expression matrix or multiple expression matrices for each batch
in a list (default:NULL)}

\item{add.batch.column}{Whether a batch column should be added to the metadata.
This is only possible if metadata is included as a list of separate dataframes
(default:TRUE).}

\item{batch.names}{Names for batches if add.batch.column==TRUE. If not set,
then the batches will be named 1,2,3,...}

\item{create.comparison.app}{Whether the bulkAnalyseR app comparing corrected
and uncorrected outputs should be created. This is only possible if uncorrected!=NULL.}

\item{comparison.app.dir}{Directory where comparison app should be saved if it is created (default: 'comparisonApp').}

\item{comparison.app.title}{Title for comparison app (default: 'Comparison between uncorrected and corrected outputs')}

\item{corrected.app.dir}{Directory where corrected app should be saved if it is created (default: 'correctedApp').}

\item{corrected.app.title}{Title for corrected app (default: 'Batch corrected bulkAnalyseR app')}

\item{organism}{organism name to be passed on to gprofiler2::gost in corrected app; organism names are constructed by concatenating the first letter of the name and the family name; default is NA - enrichment is not included to ensure compatibility with datasets that have non-standard gene names.}

\item{modality}{Name of the modality for corrected app}

\item{theme}{Shiny theme to be used in the app (default: 'flatly')}

\item{panels.default}{Argument to control which of the default panels will be included in the corrected app; default is all, but the enrichment panel will not appear unless organism is also supplied; note that the 'DE' panel is required for 'DEplot', 'DEsummary', 'Enrichment', and 'GRNenrichment'.}

\item{corrected}{Expression matrix as output by coRRectoR}

\item{org.db=}{Database for annotations to transform ENSEMBL IDs to gene names in both apps; a list of bioconductor packaged databases can be found with BiocManager::available("^org\."); default in NA, in which case the row names of the expression matrix are used directly - it is recommended to provide ENSEMBL IDs if the database for your model organism is available.}
}
\description{
This function takes a corrected expression matrix and creates a minimal
bulkAnalyseR app comparing this to the uncorrected expression matrix and a full
bulkAnalyseR app on the corrected expression matrix.
}
\examples{
# Defining corrected and uncorreced expression matrices and metadata
corrected = matrix(sample(1:100,70),ncol = 7)
colnames(corrected)=LETTERS[1:7]
uncorrected = matrix(sample(1:100,70),ncol = 7)
colnames(uncorrected)=LETTERS[1:7]

# Run preparebulkAnalyseR with combined uncorrected expression matrix 
# and split metadata
meta = list('batch1'=data.frame('id'=LETTERS[1:3],
                                'batch'=rep('old',3),
                                'condition'=c('A','A','B')),
            'batch2'=data.frame('id'=LETTERS[4:7],
                                'batch'=rep('new',4),
                                  'condition'=c('B','C','B','D')))
preparebulkAnalyseR(corrected,
                    meta,
                    uncorrected,
                    comparison.app.dir = paste0(tempdir(), "/comparisonShinyApp"),
                    corrected.app.dir = paste0(tempdir(), "/correctedShinyApp"))
# clean up tempdir
unlink(paste0(normalizePath(tempdir()), "/", dir(tempdir())), recursive = TRUE)

Run preparebulkAnalyseR with split uncorrected expression matrix 
meta = list('batch1'=data.frame('id'=LETTERS[1:3],
                                'condition'=c('A','A','B')),
            'batch2'=data.frame('id'=LETTERS[4:7],
                                'condition'=c('B','C','B','D')))
uncorrected.list = list('batch1'=uncorrected[,1:3],
                        'batch2'=uncorrected[,4:7])
preparebulkAnalyseR(corrected,
                    meta,
                    uncorrected.list,
                    batch.names=c('old','new'),
                    create.comparison.app = FALSE,
                    create.corrected.app = FALSE)
                    
# Run preparebulkAnalyseR with no uncorrected expression matrix and
# combined metadata table
preparebulkAnalyseR(corrected,
                    rbind(meta[[1]],meta[[2]]),
                    batch.names=c('old','new'),
                    create.comparison.app = FALSE,
                    create.corrected.app = FALSE)
}
